update matchslots and reorgornize files
