+++ email-calendar-backend/scheduling-engine/test.txt
@@ -1,0 +1,75 @@
+[{{kind: "calendar#event", etag: ""3111717400128000"", id: "2esvdes71u0fflekrkak9hq4on", status: "confirmed", htmlLink: "https://www.google.com/calendar/event?eid=MmVzdmRlczcxdTBmZmxla3JrYWs5aHE0b24geWlib3BhbjEyQG0", …}
+created: "2019-04-21T14:58:20.000Z"
+creator: {email: "yibopan12@gmail.com", self: true}
+end: {dateTime: "2019-04-21T10:00:00-05:00"}
+etag: ""3111717400128000""
+extendedProperties: {private: {…}}
+htmlLink: "https://www.google.com/calendar/event?eid=MmVzdmRlczcxdTBmZmxla3JrYWs5aHE0b24geWlib3BhbjEyQG0"
+iCalUID: "2esvdes71u0fflekrkak9hq4on@google.com"
+id: "2esvdes71u0fflekrkak9hq4on"
+kind: "calendar#event"
+organizer: {email: "yibopan12@gmail.com", self: true}
+reminders: {useDefault: true}
+sequence: 0
+start: {dateTime: "2019-04-21T08:00:00-05:00"}
+status: "confirmed"
+summary: "test1"
+updated: "2019-04-21T14:58:20.064Z"
+__proto__: Object
+_ProfilePageTest.js:67},
+{{kind: "calendar#event", etag: ""3111717473396000"", id: "20q1f56r4fjd9uh9m035jh2kmi", status: "confirmed", htmlLink: "https://www.google.com/calendar/event?eid=MjBxMWY1NnI0ZmpkOXVoOW0wMzVqaDJrbWkgeWlib3BhbjEyQG0", …}
+created: "2019-04-21T14:58:56.000Z"
+creator: {email: "yibopan12@gmail.com", self: true}
+end: {dateTime: "2019-04-21T14:00:00-05:00"}
+etag: ""3111717473396000""
+extendedProperties: {private: {…}}
+htmlLink: "https://www.google.com/calendar/event?eid=MjBxMWY1NnI0ZmpkOXVoOW0wMzVqaDJrbWkgeWlib3BhbjEyQG0"
+iCalUID: "20q1f56r4fjd9uh9m035jh2kmi@google.com"
+id: "20q1f56r4fjd9uh9m035jh2kmi"
+kind: "calendar#event"
+organizer: {email: "yibopan12@gmail.com", self: true}
+reminders: {useDefault: true}
+sequence: 0
+start: {dateTime: "2019-04-21T12:00:00-05:00"}
+status: "confirmed"
+summary: "test2"
+updated: "2019-04-21T14:58:56.698Z"
+__proto__: Object
+_ProfilePageTest.js:67}, 
+{{kind: "calendar#event", etag: ""3111717517350000"", id: "59isfh9qtvv7ps4sainljst0vm", status: "confirmed", htmlLink: "https://www.google.com/calendar/event?eid=NTlpc2ZoOXF0dnY3cHM0c2Fpbmxqc3Qwdm0geWlib3BhbjEyQG0", …}
+created: "2019-04-21T14:59:18.000Z"
+creator: {email: "yibopan12@gmail.com", self: true}
+end: {dateTime: "2019-04-21T16:00:00-05:00"}
+etag: ""3111717517350000""
+extendedProperties: {private: {…}}
+htmlLink: "https://www.google.com/calendar/event?eid=NTlpc2ZoOXF0dnY3cHM0c2Fpbmxqc3Qwdm0geWlib3BhbjEyQG0"
+iCalUID: "59isfh9qtvv7ps4sainljst0vm@google.com"
+id: "59isfh9qtvv7ps4sainljst0vm"
+kind: "calendar#event"
+organizer: {email: "yibopan12@gmail.com", self: true}
+reminders: {useDefault: true}
+sequence: 0
+start: {dateTime: "2019-04-21T14:30:00-05:00"}
+status: "confirmed"
+summary: "test3"
+updated: "2019-04-21T14:59:18.675Z"
+__proto__: Object
+_ProfilePageTest.js:67}, 
+{{kind: "calendar#event", etag: ""3111717602430000"", id: "5pjgkph0dosdbnc3ke0v0rqle6", status: "confirmed", htmlLink: "https://www.google.com/calendar/event?eid=NXBqZ2twaDBkb3NkYm5jM2tlMHYwcnFsZTYgeWlib3BhbjEyQG0", …}
+created: "2019-04-21T14:59:26.000Z"
+creator: {email: "yibopan12@gmail.com", self: true}
+end: {dateTime: "2019-04-21T19:12:00-05:00"}
+etag: ""3111717602430000""
+extendedProperties: {private: {…}}
+htmlLink: "https://www.google.com/calendar/event?eid=NXBqZ2twaDBkb3NkYm5jM2tlMHYwcnFsZTYgeWlib3BhbjEyQG0"
+iCalUID: "5pjgkph0dosdbnc3ke0v0rqle6@google.com"
+id: "5pjgkph0dosdbnc3ke0v0rqle6"
+kind: "calendar#event"
+organizer: {email: "yibopan12@gmail.com", self: true}
+reminders: {useDefault: true}
+sequence: 2
+start: {dateTime: "2019-04-21T17:27:00-05:00"}
+status: "confirmed"
+summary: "test4"
+updated: "2019-04-21T15:00:01.215Z"
+__proto__: Object}]
